refactor CI evidence and CLI test support
