--- crates/nose-cli/tests/cli/support.rs
@@ -1,599 +1,0 @@
-//! End-to-end CLI tests: run the built `nose` binary against a temp project and
-//! check the user-visible behavior (discovery, `query` report, `--exclude`).
-
-pub(crate) use std::fs;
-pub(crate) use std::path::{Path, PathBuf};
-use std::process::Output;
-pub(crate) use std::process::{Command, Stdio};
-use std::sync::atomic::{AtomicU64, Ordering};
-
-static TEMP_PROJECT_SEQ: AtomicU64 = AtomicU64::new(0);
-
-pub(crate) fn bin() -> &'static str {
-    env!("CARGO_BIN_EXE_nose")
-}
-
-pub(crate) fn make_temp_dir(tag: &str) -> PathBuf {
-    let unique = TEMP_PROJECT_SEQ.fetch_add(1, Ordering::Relaxed);
-    let dir = std::env::temp_dir().join(format!("nose_{tag}_{}_{}", std::process::id(), unique));
-    let _ = fs::remove_dir_all(&dir);
-    fs::create_dir_all(&dir).unwrap();
-    dir
-}
-
-pub(crate) fn write_files(dir: &Path, files: &[(&str, &str)]) {
-    for (name, src) in files {
-        let path = dir.join(name);
-        if let Some(parent) = path.parent() {
-            fs::create_dir_all(parent).unwrap();
-        }
-        fs::write(&path, src).unwrap_or_else(|err| panic!("write {}: {err}", path.display()));
-    }
-}
-
-/// Write a small project (a 3-copy clone family + a decoy) into a unique temp dir.
-pub(crate) fn make_project(tag: &str) -> PathBuf {
-    let dir = make_temp_dir(tag);
-    let body = |acc: &str, it: &str| {
-        format!("def f(items):\n    {acc} = 0\n    for {it} in items:\n        if {it} > 0:\n            {acc} = {acc} + {it} * {it}\n    return {acc}\n")
-    };
-    for (sub, src) in [
-        ("a", body("total", "x")),
-        ("b", body("acc", "v")),
-        ("tests", body("s", "n")),
-    ] {
-        let d = dir.join(sub);
-        fs::create_dir_all(&d).unwrap();
-        fs::write(d.join("f.py"), src).unwrap();
-    }
-    let d = dir.join("c");
-    fs::create_dir_all(&d).unwrap();
-    fs::write(
-        d.join("decoy.py"),
-        "def greet(n):\n    m = 'hi ' + n\n    print(m)\n    print(n)\n    return m\n",
-    )
-    .unwrap();
-    dir
-}
-
-pub(crate) fn make_mode_project(tag: &str) -> PathBuf {
-    let dir = make_temp_dir(tag);
-
-    fs::write(
-        dir.join("renamed_a.py"),
-        "def total(items):\n    total = 0\n    for item in items:\n        if item > 0:\n            total = total + item * item\n    return total\n",
-    )
-    .unwrap();
-    fs::write(
-        dir.join("renamed_b.py"),
-        "def score(values):\n    acc = 0\n    for value in values:\n        if value > 0:\n            acc = acc + value * value\n    return acc\n",
-    )
-    .unwrap();
-
-    let copied = "def copied(events):\n    out = []\n    for e in events:\n        if e.kind == 1:\n            out.append(e.payload)\n            record(e.id, e.kind)\n    return out\n";
-    fs::write(dir.join("copy_a.py"), copied).unwrap();
-    fs::write(dir.join("copy_b.py"), copied).unwrap();
-
-    // A near-duplicate pair: mostly the same body, but each side has one
-    // statement the other lacks — different raw tokens (no syntax run),
-    // different value fingerprints (no exact semantic), near-similar shape.
-    fs::write(
-        dir.join("near_a.py"),
-        "def fetch_users(client, ids):\n    results = []\n    for uid in ids:\n        row = client.get(uid)\n        if row is None:\n            continue\n        results.append(normalize(row))\n    log_count(len(results))\n    return results\n",
-    )
-    .unwrap();
-    fs::write(
-        dir.join("near_b.py"),
-        "def fetch_orders(client, keys):\n    found = []\n    for key in keys:\n        record = client.get(key)\n        if record is None:\n            continue\n        found.append(normalize(record))\n        audit(record)\n    return found\n",
-    )
-    .unwrap();
-
-    dir
-}
-
-pub(crate) fn make_fragment_project(tag: &str) -> PathBuf {
-    let dir = make_temp_dir(tag);
-    fs::create_dir_all(dir.join("a")).unwrap();
-    fs::create_dir_all(dir.join("b")).unwrap();
-    fs::write(
-        dir.join("a/f.py"),
-        "def first(xs):\n    total = 0\n    if xs:\n        return xs[0] + 1\n    return total\n",
-    )
-    .unwrap();
-    fs::write(
-        dir.join("b/f.py"),
-        "def second(items):\n    acc = 10\n    if items:\n        return items[0] + 1\n    return acc\n",
-    )
-    .unwrap();
-    dir
-}
-
-pub(crate) fn make_generated_header_project(tag: &str) -> PathBuf {
-    let dir = make_temp_dir(tag);
-    let body = |name: &str, acc: &str, it: &str| {
-        format!(
-            "# Generated by fixture\n\
-def {name}(items):\n    {acc} = 0\n    for {it} in items:\n        if {it} > 0:\n            {acc} = {acc} + {it} * {it}\n    return {acc}\n"
-        )
-    };
-    for (sub, src) in [
-        ("a", body("first", "total", "x")),
-        ("b", body("second", "acc", "v")),
-        ("c", body("third", "sum", "n")),
-    ] {
-        let d = dir.join(sub);
-        fs::create_dir_all(&d).unwrap();
-        fs::write(d.join("f.py"), src).unwrap();
-    }
-    dir
-}
-
-pub(crate) fn make_jazzy_generated_project(tag: &str) -> PathBuf {
-    let dir = make_temp_dir(tag);
-    let body = |title: &str, name: &str, acc: &str, item: &str| {
-        format!(
-            "<!doctype html>\n<html><head>\n{}\n<link rel=\"stylesheet\" href=\"../css/jazzy.css\">\n</head><body>\n<a class=\"dashAnchor\" name=\"//apple_ref/swift/cl/{title}\"></a>\n<section class=\"content\"><div class=\"task-group\"><h2>Methods</h2></div></section>\n<script>\nfunction {name}(items) {{\n  let {acc} = 0;\n  for (const {item} of items) {{\n    if ({item} > 0) {{\n      {acc} = {acc} + {item} * {item};\n    }}\n  }}\n  return {acc};\n}}\n</script>\n</body></html>\n",
-            "<meta name=\"generator-padding\" content=\"ordinary\">\n".repeat(10)
-        )
-    };
-    for (sub, src) in [
-        ("a", body("First", "first", "total", "value")),
-        ("b", body("Second", "second", "sum", "item")),
-        ("c", body("Third", "third", "acc", "entry")),
-    ] {
-        let child = dir.join(sub);
-        fs::create_dir_all(&child).unwrap();
-        fs::write(child.join("index.html"), src).unwrap();
-    }
-    dir
-}
-
-pub(crate) fn make_declared_generator_project(tag: &str) -> PathBuf {
-    let dir = make_temp_dir(tag);
-    let body = |title: &str, name: &str, acc: &str, item: &str| {
-        format!(
-            "<!doctype html>\n<html><head>\n<meta content=\"Static Site Tool 4.2\" name=\"generator\">\n<title>{title}</title>\n</head><body>\n<script>\nfunction {name}(items) {{\n  let {acc} = 0;\n  for (const {item} of items) {{\n    if ({item} > 0) {{\n      {acc} = {acc} + {item} * {item};\n    }}\n  }}\n  return {acc};\n}}\n</script>\n</body></html>\n"
-        )
-    };
-    for (sub, src) in [
-        ("a", body("First", "first", "total", "value")),
-        ("b", body("Second", "second", "sum", "item")),
-        ("c", body("Third", "third", "acc", "entry")),
-    ] {
-        let child = dir.join(sub);
-        fs::create_dir_all(&child).unwrap();
-        fs::write(child.join("index.html"), src).unwrap();
-    }
-    dir
-}
-
-pub(crate) struct TempProject {
-    dir: PathBuf,
-}
-
-impl TempProject {
-    pub(crate) fn new(tag: &str) -> Self {
-        Self {
-            dir: make_temp_dir(tag),
-        }
-    }
-
-    pub(crate) fn write(&self, name: &str, src: &str) {
-        write_files(&self.dir, &[(name, src)]);
-    }
-
-    pub(crate) fn path(&self) -> &Path {
-        &self.dir
-    }
-
-    pub(crate) fn query_json(&self, mode: &str, extra_args: &[&str]) -> serde_json::Value {
-        let mut args = vec!["query", self.dir.to_str().unwrap(), "--mode", mode];
-        args.extend_from_slice(extra_args);
-        args.extend_from_slice(&["--format", "json"]);
-        query_json(&run(&args))
-    }
-
-    pub(crate) fn query_semantic_json(&self) -> serde_json::Value {
-        self.query_json("semantic", &[])
-    }
-
-    pub(crate) fn query_semantic_min_json(&self) -> serde_json::Value {
-        self.query_json(
-            "semantic",
-            &["top=0", "--min-size", "1", "--min-lines", "1"],
-        )
-    }
-}
-
-impl Drop for TempProject {
-    fn drop(&mut self) {
-        let _ = fs::remove_dir_all(&self.dir);
-    }
-}
-
-fn query_json_list_args(args: &[&str]) -> Vec<String> {
-    let mut normalized = args
-        .iter()
-        .map(|arg| (*arg).to_string())
-        .collect::<Vec<_>>();
-    if args.first() != Some(&"query") || !args.windows(2).any(|w| w == ["--format", "json"]) {
-        return normalized;
-    }
-    let mut has_term = false;
-    let mut has_top = false;
-    let mut skip_next = false;
-    for arg in args.iter().skip(2) {
-        if skip_next {
-            skip_next = false;
-            continue;
-        }
-        if matches!(
-            *arg,
-            "--mode"
-                | "--min-size"
-                | "--min-lines"
-                | "--min-value"
-                | "--min-members"
-                | "--format"
-                | "--baseline"
-                | "--ignore-file"
-                | "--config"
-                | "--cache-dir"
-                | "--semantic-pack"
-                | "--exclude"
-                | "--generated-path"
-                | "--fail-on"
-        ) {
-            skip_next = true;
-            continue;
-        }
-        if !arg.starts_with('-') {
-            if arg.starts_with("top=") {
-                has_top = true;
-                continue;
-            }
-            has_term = true;
-            break;
-        }
-    }
-    if !has_term {
-        if !has_top {
-            normalized.insert(2, "top=0".to_string());
-        }
-        normalized.insert(2, "all".to_string());
-    }
-    normalized
-}
-
-fn owned_args(args: &[&str]) -> Vec<String> {
-    args.iter().map(|arg| (*arg).to_string()).collect()
-}
-
-pub(crate) fn run(args: &[&str]) -> String {
-    let normalized = query_json_list_args(args);
-    let out = nose_output(&normalized);
-    assert_success(&normalized, &out);
-    String::from_utf8(out.stdout).unwrap()
-}
-
-pub(crate) fn run_raw(args: &[&str]) -> String {
-    let args = owned_args(args);
-    let out = nose_output(&args);
-    assert_success(&args, &out);
-    String::from_utf8(out.stdout).unwrap()
-}
-
-/// Like [`run`] but expects a non-zero exit; returns stderr (where errors print).
-#[allow(dead_code)]
-pub(crate) fn run_fail(args: &[&str]) -> String {
-    let args = owned_args(args);
-    let out = nose_output(&args);
-    assert!(
-        !out.status.success(),
-        "{}",
-        cli_output_message_with_bin(
-            bin(),
-            "expected nose to fail, but it succeeded",
-            &args,
-            &format!("{:?}", out.status),
-            &out.stdout,
-            &out.stderr
-        )
-    );
-    String::from_utf8(out.stderr).unwrap()
-}
-
-fn nose_output(args: &[String]) -> Output {
-    Command::new(bin())
-        .args(args)
-        .output()
-        .unwrap_or_else(|err| {
-            panic!(
-                "failed to spawn nose\ncommand: {}\nerror: {err}",
-                command_line_for_message(bin(), args)
-            )
-        })
-}
-
-fn assert_success(args: &[String], out: &Output) {
-    assert!(
-        out.status.success(),
-        "{}",
-        cli_output_message_with_bin(
-            bin(),
-            "nose exited non-zero",
-            args,
-            &format!("{:?}", out.status),
-            &out.stdout,
-            &out.stderr
-        )
-    );
-}
-
-fn cli_output_message_with_bin(
-    bin: &str,
-    summary: &str,
-    args: &[String],
-    status: &str,
-    stdout: &[u8],
-    stderr: &[u8],
-) -> String {
-    format!(
-        "{summary}\nstatus: {status}\ncommand: {}\nstdout:\n{}\nstderr:\n{}",
-        command_line_for_message(bin, args),
-        String::from_utf8_lossy(stdout),
-        String::from_utf8_lossy(stderr)
-    )
-}
-
-fn command_line_for_message(bin: &str, args: &[String]) -> String {
-    let mut parts = Vec::with_capacity(args.len() + 1);
-    parts.push(bin.to_string());
-    parts.extend(args.iter().map(|arg| format!("{arg:?}")));
-    parts.join(" ")
-}
-
-/// `nose query <dir> --mode <mode> --format json --top 0` with the tiny-fixture
-/// floors (`--min-lines 1 --min-size 1`) — the standard invocation for the
-/// one-function fixtures the semantic CLI tests are built from.
-pub(crate) fn query_min_json(dir: &Path, mode: &str) -> String {
-    run(&[
-        "query",
-        dir.to_str().unwrap(),
-        "--mode",
-        mode,
-        "--min-lines",
-        "1",
-        "--min-size",
-        "1",
-        "--format",
-        "json",
-        "top=0",
-    ])
-}
-
-pub(crate) fn add_distinct_clone_family(dir: &Path) {
-    let d = dir.join("new");
-    fs::create_dir_all(&d).unwrap();
-    let body = |name: &str, acc: &str, it: &str| {
-        format!(
-            "def {name}(items):\n    {acc} = 1\n    for {it} in items:\n        if {it} < 10:\n            {acc} = {acc} * ({it} + 3)\n            {acc} = {acc} - {it}\n    return {acc}\n"
-        )
-    };
-    fs::write(d.join("fresh_a.py"), body("fresh_a", "total", "item")).unwrap();
-    fs::write(d.join("fresh_b.py"), body("fresh_b", "score", "value")).unwrap();
-}
-
-pub(crate) fn add_member_to_existing_family(dir: &Path) {
-    let d = dir.join("d");
-    fs::create_dir_all(&d).unwrap();
-    fs::write(
-        d.join("f.py"),
-        "def f(items):\n    sum = 0\n    for z in items:\n        if z > 0:\n            sum = sum + z * z\n    return sum\n",
-    )
-    .unwrap();
-}
-
-pub(crate) fn query_json(out: &str) -> serde_json::Value {
-    normalize_query_json(serde_json::from_str(out).expect("query should emit valid JSON"))
-}
-
-fn normalize_query_json(mut json: serde_json::Value) -> serde_json::Value {
-    if let Some(families) = json["families"].as_array_mut() {
-        for family in families {
-            if family["family_id"].is_null() && !family["id"].is_null() {
-                family["family_id"] = family["id"].clone();
-            }
-            if family["recommended_surface"].is_null() && !family["surface"].is_null() {
-                family["recommended_surface"] = family["surface"].clone();
-            }
-            if family["shared_lines"].is_null() && !family["shared"].is_null() {
-                family["shared_lines"] = family["shared"].clone();
-            }
-            if family["dup_lines"].is_null() && !family["removable"].is_null() {
-                family["dup_lines"] = family["removable"].clone();
-            }
-            if family["modules"].is_null() && !family["dirs"].is_null() {
-                family["modules"] = family["dirs"].clone();
-            }
-            if let Some(locations) = family["locations"].as_array_mut() {
-                for loc in locations {
-                    if loc["start_line"].is_null() && !loc["start"].is_null() {
-                        loc["start_line"] = loc["start"].clone();
-                    }
-                    if loc["end_line"].is_null() && !loc["end"].is_null() {
-                        loc["end_line"] = loc["end"].clone();
-                    }
-                    if loc["kind"].is_null() {
-                        loc["kind"] = serde_json::Value::from("Block");
-                    }
-                    if loc["fragment_kind"].is_null() {
-                        loc["fragment_kind"] = serde_json::Value::from("conditional-guard");
-                    }
-                    if loc["reason_code"].is_null() {
-                        loc["reason_code"] = serde_json::Value::from("exact-conditional-guard");
-                    }
-                    if loc["is_fragment"].is_null() {
-                        loc["is_fragment"] = serde_json::Value::from(true);
-                    }
-                    if loc["span_lines"].is_null() {
-                        let start = loc["start_line"].as_u64().unwrap_or(0);
-                        let end = loc["end_line"].as_u64().unwrap_or(start);
-                        loc["span_lines"] = serde_json::Value::from(end.saturating_sub(start) + 1);
-                    }
-                }
-            }
-        }
-    }
-    json
-}
-
-pub(crate) fn query_families(json: &serde_json::Value) -> &[serde_json::Value] {
-    json["families"]
-        .as_array()
-        .expect("query JSON should contain families array")
-}
-
-pub(crate) fn family_contains_all(json: &serde_json::Value, suffixes: &[&str]) -> bool {
-    family_with_all(json, suffixes).is_some()
-}
-
-pub(crate) fn family_with_all<'a>(
-    json: &'a serde_json::Value,
-    suffixes: &[&str],
-) -> Option<&'a serde_json::Value> {
-    query_families(json).iter().find(|family| {
-        let Some(locations) = family["locations"].as_array() else {
-            return false;
-        };
-        suffixes.iter().all(|suffix| {
-            locations.iter().any(|loc| {
-                loc["file"]
-                    .as_str()
-                    .is_some_and(|file| file.ends_with(suffix))
-            })
-        })
-    })
-}
-
-/// Find a family that pairs `left` and `right` as locations at exactly
-/// `start_line..end_line`, where every location is a `Block` and none sits in
-/// `negative` — the standard positive check for branch-boundary fragment tests.
-pub(crate) fn block_branch_pair_family<'a>(
-    families: &'a [serde_json::Value],
-    left: &str,
-    right: &str,
-    negative: &str,
-    start_line: u64,
-    end_line: u64,
-) -> Option<&'a serde_json::Value> {
-    families.iter().find(|family| {
-        let locations = family["locations"].as_array().expect("locations");
-        let branch_files: Vec<&str> = locations
-            .iter()
-            .filter(|loc| loc["start_line"] == start_line && loc["end_line"] == end_line)
-            .filter_map(|loc| loc["file"].as_str())
-            .collect();
-        branch_files.iter().any(|file| file.ends_with(left))
-            && branch_files.iter().any(|file| file.ends_with(right))
-            && locations.iter().all(|loc| loc["kind"] == "Block")
-            && locations
-                .iter()
-                .filter_map(|loc| loc["file"].as_str())
-                .all(|file| !file.ends_with(negative))
-    })
-}
-
-/// Whether any family pairs `left` at `(start, end)` with `right` at its
-/// `(start, end)` — the standard negative check for branch-boundary tests
-/// (asserting two spans were NOT merged into one family).
-pub(crate) fn families_pair_locations(
-    families: &[serde_json::Value],
-    left: (&str, u64, u64),
-    right: (&str, u64, u64),
-) -> bool {
-    fn has_location(locations: &[serde_json::Value], (file, start, end): (&str, u64, u64)) -> bool {
-        locations.iter().any(|loc| {
-            loc["file"].as_str().is_some_and(|f| f.ends_with(file))
-                && loc["start_line"] == start
-                && loc["end_line"] == end
-        })
-    }
-    families.iter().any(|family| {
-        let locations = family["locations"].as_array().expect("locations");
-        has_location(locations, left) && has_location(locations, right)
-    })
-}
-
-pub(crate) fn json_array_strings<'a>(value: &'a serde_json::Value, key: &str) -> Vec<&'a str> {
-    value[key]
-        .as_array()
-        .unwrap_or_else(|| panic!("{key} should be an array"))
-        .iter()
-        .map(|item| {
-            item.as_str()
-                .unwrap_or_else(|| panic!("{key} entries should be strings"))
-        })
-        .collect()
-}
-
-pub(crate) fn assert_query_json_reports_semantic_packs(json: &serde_json::Value) {
-    let packs = json["semantic_packs"]
-        .as_array()
-        .unwrap_or_else(|| panic!("query JSON should report semantic_packs: {json}"));
-    for id in [
-        "nose.first_party",
-        "nose.lang.python",
-        "nose.lang.javascript-typescript",
-        "nose.lang.go",
-        "nose.lang.rust",
-        "nose.lang.java",
-        "nose.lang.c",
-        "nose.lang.ruby",
-        "nose.lang.swift",
-        "nose.lang.css",
-        "nose.lang.html",
-        "nose.python.builtins.collection_factories",
-        "nose.python.stdlib.collection_factories",
-        "nose.python.stdlib.math",
-        "nose.ruby.stdlib.set",
-        "nose.rust.stdlib.vec",
-        "nose.rust.stdlib.option",
-        "nose.rust.stdlib.integer_methods",
-        "nose.rust.stdlib.collection_factories",
-        "nose.rust.stdlib.map_factories",
-        "nose.java.stdlib.math",
-        "nose.java.stdlib.map_factories",
-        "nose.java.stdlib.map_entries",
-        "nose.java.stdlib.collection_factories",
-        "nose.java.stdlib.collection_constructors",
-        "nose.java.stdlib.static_collection_adapters",
-        "nose.protocols.map_get",
-        "nose.protocols.map_get_default",
-        "nose.protocols.free_function_builtins",
-        "nose.protocols.iterator_builtins",
-        "nose.protocols.receiver_membership",
-        "nose.protocols.map_key_views",
-        "nose.protocols.property_builtins",
-        "nose.protocols.builtin_method_calls",
-        "nose.go.stdlib.namespace_calls",
-        "nose.protocols.iterator_identity_adapters",
-        "nose.javascript.builtins.promise",
-        "nose.javascript.builtins.array",
-        "nose.javascript.builtins.boolean",
-        "nose.javascript.builtins.regex",
-        "nose.javascript.builtins.static_index_membership",
-        "nose.javascript.builtins.collection_constructors",
-        "nose.python.stdlib.type_domain",
-        "nose.value_graph.laws",
-    ] {
-        assert!(
-            packs.iter().any(|pack| pack["id"] == id),
-            "query JSON should report builtin semantic pack {id}: {json}"
-        );
-    }
-}
-
-#[cfg(test)]
-#[path = "support_tests.rs"]
-mod tests;
