+++ crates/nose-cli/tests/cli/support/fixtures.rs
@@ -1,0 +1,210 @@
+use std::fs;
+use std::path::{Path, PathBuf};
+use std::sync::atomic::{AtomicU64, Ordering};
+
+use super::process::run;
+use super::query::query_json;
+
+static TEMP_PROJECT_SEQ: AtomicU64 = AtomicU64::new(0);
+
+pub(crate) fn bin() -> &'static str {
+    env!("CARGO_BIN_EXE_nose")
+}
+
+pub(crate) fn make_temp_dir(tag: &str) -> PathBuf {
+    let unique = TEMP_PROJECT_SEQ.fetch_add(1, Ordering::Relaxed);
+    let dir = std::env::temp_dir().join(format!("nose_{tag}_{}_{}", std::process::id(), unique));
+    let _ = fs::remove_dir_all(&dir);
+    fs::create_dir_all(&dir).unwrap();
+    dir
+}
+
+pub(crate) fn write_files(dir: &Path, files: &[(&str, &str)]) {
+    for (name, src) in files {
+        let path = dir.join(name);
+        if let Some(parent) = path.parent() {
+            fs::create_dir_all(parent).unwrap();
+        }
+        fs::write(&path, src).unwrap_or_else(|err| panic!("write {}: {err}", path.display()));
+    }
+}
+
+/// Write a small project (a 3-copy clone family + a decoy) into a unique temp dir.
+pub(crate) fn make_project(tag: &str) -> PathBuf {
+    let dir = make_temp_dir(tag);
+    let body = |acc: &str, it: &str| {
+        format!("def f(items):\n    {acc} = 0\n    for {it} in items:\n        if {it} > 0:\n            {acc} = {acc} + {it} * {it}\n    return {acc}\n")
+    };
+    for (sub, src) in [
+        ("a", body("total", "x")),
+        ("b", body("acc", "v")),
+        ("tests", body("s", "n")),
+    ] {
+        let d = dir.join(sub);
+        fs::create_dir_all(&d).unwrap();
+        fs::write(d.join("f.py"), src).unwrap();
+    }
+    let d = dir.join("c");
+    fs::create_dir_all(&d).unwrap();
+    fs::write(
+        d.join("decoy.py"),
+        "def greet(n):\n    m = 'hi ' + n\n    print(m)\n    print(n)\n    return m\n",
+    )
+    .unwrap();
+    dir
+}
+
+pub(crate) fn make_mode_project(tag: &str) -> PathBuf {
+    let dir = make_temp_dir(tag);
+
+    fs::write(
+        dir.join("renamed_a.py"),
+        "def total(items):\n    total = 0\n    for item in items:\n        if item > 0:\n            total = total + item * item\n    return total\n",
+    )
+    .unwrap();
+    fs::write(
+        dir.join("renamed_b.py"),
+        "def score(values):\n    acc = 0\n    for value in values:\n        if value > 0:\n            acc = acc + value * value\n    return acc\n",
+    )
+    .unwrap();
+
+    let copied = "def copied(events):\n    out = []\n    for e in events:\n        if e.kind == 1:\n            out.append(e.payload)\n            record(e.id, e.kind)\n    return out\n";
+    fs::write(dir.join("copy_a.py"), copied).unwrap();
+    fs::write(dir.join("copy_b.py"), copied).unwrap();
+
+    // A near-duplicate pair: mostly the same body, but each side has one
+    // statement the other lacks — different raw tokens (no syntax run),
+    // different value fingerprints (no exact semantic), near-similar shape.
+    fs::write(
+        dir.join("near_a.py"),
+        "def fetch_users(client, ids):\n    results = []\n    for uid in ids:\n        row = client.get(uid)\n        if row is None:\n            continue\n        results.append(normalize(row))\n    log_count(len(results))\n    return results\n",
+    )
+    .unwrap();
+    fs::write(
+        dir.join("near_b.py"),
+        "def fetch_orders(client, keys):\n    found = []\n    for key in keys:\n        record = client.get(key)\n        if record is None:\n            continue\n        found.append(normalize(record))\n        audit(record)\n    return found\n",
+    )
+    .unwrap();
+
+    dir
+}
+
+pub(crate) fn make_fragment_project(tag: &str) -> PathBuf {
+    let dir = make_temp_dir(tag);
+    fs::create_dir_all(dir.join("a")).unwrap();
+    fs::create_dir_all(dir.join("b")).unwrap();
+    fs::write(
+        dir.join("a/f.py"),
+        "def first(xs):\n    total = 0\n    if xs:\n        return xs[0] + 1\n    return total\n",
+    )
+    .unwrap();
+    fs::write(
+        dir.join("b/f.py"),
+        "def second(items):\n    acc = 10\n    if items:\n        return items[0] + 1\n    return acc\n",
+    )
+    .unwrap();
+    dir
+}
+
+pub(crate) fn make_generated_header_project(tag: &str) -> PathBuf {
+    let dir = make_temp_dir(tag);
+    let body = |name: &str, acc: &str, it: &str| {
+        format!(
+            "# Generated by fixture\n\
+def {name}(items):\n    {acc} = 0\n    for {it} in items:\n        if {it} > 0:\n            {acc} = {acc} + {it} * {it}\n    return {acc}\n"
+        )
+    };
+    for (sub, src) in [
+        ("a", body("first", "total", "x")),
+        ("b", body("second", "acc", "v")),
+        ("c", body("third", "sum", "n")),
+    ] {
+        let d = dir.join(sub);
+        fs::create_dir_all(&d).unwrap();
+        fs::write(d.join("f.py"), src).unwrap();
+    }
+    dir
+}
+
+pub(crate) fn make_jazzy_generated_project(tag: &str) -> PathBuf {
+    let dir = make_temp_dir(tag);
+    let body = |title: &str, name: &str, acc: &str, item: &str| {
+        format!(
+            "<!doctype html>\n<html><head>\n{}\n<link rel=\"stylesheet\" href=\"../css/jazzy.css\">\n</head><body>\n<a class=\"dashAnchor\" name=\"//apple_ref/swift/cl/{title}\"></a>\n<section class=\"content\"><div class=\"task-group\"><h2>Methods</h2></div></section>\n<script>\nfunction {name}(items) {{\n  let {acc} = 0;\n  for (const {item} of items) {{\n    if ({item} > 0) {{\n      {acc} = {acc} + {item} * {item};\n    }}\n  }}\n  return {acc};\n}}\n</script>\n</body></html>\n",
+            "<meta name=\"generator-padding\" content=\"ordinary\">\n".repeat(10)
+        )
+    };
+    for (sub, src) in [
+        ("a", body("First", "first", "total", "value")),
+        ("b", body("Second", "second", "sum", "item")),
+        ("c", body("Third", "third", "acc", "entry")),
+    ] {
+        let child = dir.join(sub);
+        fs::create_dir_all(&child).unwrap();
+        fs::write(child.join("index.html"), src).unwrap();
+    }
+    dir
+}
+
+pub(crate) fn make_declared_generator_project(tag: &str) -> PathBuf {
+    let dir = make_temp_dir(tag);
+    let body = |title: &str, name: &str, acc: &str, item: &str| {
+        format!(
+            "<!doctype html>\n<html><head>\n<meta content=\"Static Site Tool 4.2\" name=\"generator\">\n<title>{title}</title>\n</head><body>\n<script>\nfunction {name}(items) {{\n  let {acc} = 0;\n  for (const {item} of items) {{\n    if ({item} > 0) {{\n      {acc} = {acc} + {item} * {item};\n    }}\n  }}\n  return {acc};\n}}\n</script>\n</body></html>\n"
+        )
+    };
+    for (sub, src) in [
+        ("a", body("First", "first", "total", "value")),
+        ("b", body("Second", "second", "sum", "item")),
+        ("c", body("Third", "third", "acc", "entry")),
+    ] {
+        let child = dir.join(sub);
+        fs::create_dir_all(&child).unwrap();
+        fs::write(child.join("index.html"), src).unwrap();
+    }
+    dir
+}
+
+pub(crate) struct TempProject {
+    dir: PathBuf,
+}
+
+impl TempProject {
+    pub(crate) fn new(tag: &str) -> Self {
+        Self {
+            dir: make_temp_dir(tag),
+        }
+    }
+
+    pub(crate) fn write(&self, name: &str, src: &str) {
+        write_files(&self.dir, &[(name, src)]);
+    }
+
+    pub(crate) fn path(&self) -> &Path {
+        &self.dir
+    }
+
+    pub(crate) fn query_json(&self, mode: &str, extra_args: &[&str]) -> serde_json::Value {
+        let mut args = vec!["query", self.dir.to_str().unwrap(), "--mode", mode];
+        args.extend_from_slice(extra_args);
+        args.extend_from_slice(&["--format", "json"]);
+        query_json(&run(&args))
+    }
+
+    pub(crate) fn query_semantic_json(&self) -> serde_json::Value {
+        self.query_json("semantic", &[])
+    }
+
+    pub(crate) fn query_semantic_min_json(&self) -> serde_json::Value {
+        self.query_json(
+            "semantic",
+            &["top=0", "--min-size", "1", "--min-lines", "1"],
+        )
+    }
+}
+
+impl Drop for TempProject {
+    fn drop(&mut self) {
+        let _ = fs::remove_dir_all(&self.dir);
+    }
+}
